generic-cfg-builder: Move the statement handlers into this file
They very well belong there, and it makes reading the code easier.
