+++ src/control-flow/generic-cfg-builder.ts
@@ -6,10 +6,32 @@ import {
   type BuilderOptions,
   type CFG,
 } from "./cfg-defs";
-import type { StatementHandlers } from "./statement-handlers";
 import { BlockMatcher } from "./block-matcher";
 import { NodeMapper } from "./node-mapper";
 import { pairwise } from "./zip";
+
+interface Dispatch {
+  single(syntax: Parser.SyntaxNode | null): BasicBlock;
+  many(statements: Parser.SyntaxNode[]): BasicBlock;
+}
+interface Link {
+  syntaxToNode: InstanceType<typeof NodeMapper>["linkSytaxToNode"];
+  offsetToSyntax: InstanceType<typeof NodeMapper>["linkOffsetToSyntax"];
+}
+export interface Context {
+  builder: Builder;
+  options: BuilderOptions;
+  matcher: BlockMatcher;
+  dispatch: Dispatch;
+  state: BlockHandler;
+  link: Link;
+}
+
+type StatementHandler = (syntax: Parser.SyntaxNode, ctx: Context) => BasicBlock;
+export type StatementHandlers = {
+  named: { [key: string]: StatementHandler };
+  default: StatementHandler;
+};
 
 export class GenericCFGBuilder {
   private builder: Builder = new Builder();
